requirements.txt
# ====================================
# Backend Dependencies
# ====================================

# Telegram Bot Framework
aiogram==3.13.1
aiogram-dialog==2.2.0  # Для более сложных диалогов (опционально)

# Web Framework (для Mini App API)
fastapi==0.115.0
uvicorn[standard]==0.30.6  # ASGI сервер
pydantic==2.9.0  # Валидация данных
pydantic-settings==2.5.0  # Настройки из .env

# Database
sqlalchemy==2.0.36  # ORM
alembic==1.13.3  # Миграции
asyncpg==0.29.0  # PostgreSQL драйвер (async)
psycopg2-binary==2.9.9  # PostgreSQL драйвер (sync, для Alembic)

# Redis (кеш, очереди) - опционально для MVP
redis==5.0.8
aioredis==2.0.1

# HTTP клиент (для внешних API)
httpx==0.27.0
aiohttp==3.10.5

# Работа с датой и временем
python-dateutil==2.9.0

# Работа с переменными окружения
python-dotenv==1.0.1

# Логирование
loguru==0.7.2  # Удобное логирование (альтернатива стандартному logging)

# Валидация и безопасность
cryptography==43.0.1  # Для проверки Telegram initData
pyjwt==2.9.0  # JWT токены (если нужны)

# Работа с файлами
pillow==10.4.0  # Обработка изображений (для будущих фич)
python-multipart==0.0.9  # Для загрузки файлов в FastAPI

# Оплата (ЮKassa)
yookassa==3.0.0

# Генерация PDF (для сертификатов в Full версии)
# reportlab==4.0.5
# weasyprint==60.1

# Тестирование (опционально)
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0  # Для тестирования FastAPI

# Форматирование кода (опционально)
black==24.8.0
flake8==7.1.1
mypy==1.11.2

# ====================================
# Production (для деплоя)
# ====================================

# ASGI сервер для продакшена
gunicorn==23.0.0

# Monitoring (опционально)
# sentry-sdk==2.6.0  # Отслеживание ошибок

